requirements.txt
streamlit
yfinance
google-generativeai
pandas
plotly
python-dotenv
